authentication added

+++ graphql/resolvers/booking.js
@@ -3,7 +3,10 @@ const Event = require('../../models/events')
 const {transformEvent, transformBooking} = require("./global_resolvers");
 
 module.exports = {
-    bookings: async () => {
+    bookings: async (req) => {
+        if(!req.isAuth){
+            throw new Error("Unauthenticated")
+        }
         try {
             const bookings = await Booking.find();
             return bookings.map(booking => {
@@ -13,16 +16,22 @@ module.exports = {
             throw error
         }
     },
-    createBooking: async (args)=>{
+    createBooking: async (args,req)=>{
+        if(!req.isAuth){
+            throw new Error("Unauthenticated")
+        }
        const fetchedEvent = await Event.findOne({_id:args.eventId})
        const booking = new Booking({
-           user:"61b226700bdc068facf4a291",
+           user:req.userId,
            event:fetchedEvent
        })
        const result = await booking.save();
        return transformBooking(result);
     },
-    cancelBooking: async (args)=>{
+    cancelBooking: async (args,req)=>{
+        if(!req.isAuth){
+            throw new Error("Unauthenticated")
+        }
         const booking = await Booking.findById(args.bookingId).populate('event');
         const event =  transformEvent(booking.event);
         await Booking.deleteOne({_id:args.bookingId});
